% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_tools.R
\name{resize_image}
\alias{resize_image}
\title{resize a image to match maximum width and heigh and keep ratio}
\usage{
resize_image(img_path, max_width = 400, max_height = 300)
}
\arguments{
\item{img_path}{A file path to the input image file}

\item{max_width}{A desired maximum image width, default 400 pix.}

\item{max_height}{A desired maximum image height, default 300 pix.}
}
\value{
a temporary file path to resized image.
}
\description{
resize a image to match maximum width and heigh and keep ratio
}
\examples{
# img_path <- './image/example.jpg'
# resized_img_path <- resize_image(img_path, max_width=640, max_heigh=480)
}
