% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmapi.R
\name{set_illm_conn}
\alias{set_illm_conn}
\title{interal function to Set connection to large language model}
\usage{
set_illm_conn(url = NULL, key = NULL, timeout_seconds = 30)
}
\arguments{
\item{url}{the full url of model}

\item{key}{the api_key of the large language model. It has to be obtain from the model webpage.}

\item{timeout_seconds}{, default value is 30}
}
\value{
a htt2r request object with model server url and api key.
}
\description{
interal function to Set connection to large language model
}
\examples{
 conn <- set_gllm_conn(url = "https://openrouter.ai/api/v1/chat/completions",
                       api_key = 'sk_fake_key')

}
