% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmapi.R
\name{set_llm_conn}
\alias{set_llm_conn}
\title{Set connection to openrouter model api service}
\usage{
set_llm_conn()
}
\value{
a htt2r request object. which connect openrouter
}
\description{
Set connection to openrouter model api service
}
\examples{
 conn <- set_llm_conn()

}
