% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmapi.R
\name{set_llm_conn}
\alias{set_llm_conn}
\title{Set connection to openrouter model api service}
\usage{
set_llm_conn(api_server = NULL, api_key = NULL)
}
\value{
a htt2r request object.
        default value will be  a connection instance of openrouter url and api_key
}
\description{
Set connection to openrouter model api service
}
\examples{
 conn <- set_llm_conn()

}
